Fixed -Wadssign warnings and added Wassign to check for next time

+++ checkfocus.h
@@ -41,7 +41,7 @@ extern int verbose;
  */
 
 typedef int Trow;
-typedef unsigned long Tcol;
+typedef int Tcol;
